Build with the lambda.norpc build tag
https://www.capitalone.com/tech/cloud/custom-runtime-for-go-based-lambda-functions/

Apparently the go1.x runtime for Lambda uses some kind of custom net/rpc
magic to get events to the handler, and for functions that only need the
newer custom runtime support there's an option to drop this legacy code
and shave a little bit of size off the final binary. For my typical
projects, we're talking a couple hundred kB at most out of 17 - 18 MB
binaries, so I certainly don't want to oversell the impact. But since
hfc flat-out does not support the legacy runtime we might as well take
whatever wins we can get.