Implement deployment of Lambda packages
It is now possible to deploy CloudFormation stacks using S3 code URIs
instead of container image references.

Since changing the package type of a Lambda function requires complete
replacement, this isn't great for anything that uses a Lambda function
URL and wants to have the chance to migrate smoothly. So I'm thinking
that image and S3 uploads should no longer be mutually exclusive, and if
the app is configured for both hfc should just upload both.