Allow specifying a prefix for S3 keys
This enables multiple apps to keep their code in the same S3 bucket, to
reduce the number of S3 buckets that a hypothetical hfc user would need
to manage.

It's not like my personal AWS account is anywhere near the 1,000 bucket
limit, but in my professional life the bucket limit is sadly an issue
near and dear to my heart.