Clean up code for getting stack and S3 data
This refactoring will make it easier to reuse the logic for getting the
current S3 key from every deployed stack, which I want for an `hfc
status` command that summarizes the code versions of all known stacks.

This also gets rid of the errgroup pattern where I had a closure writing
results via a pointer, and replaces it with my more standard approach of
naming the error parameter in the `group.Go` function and assigning to
differently-scoped variables with a bare return. The other pattern was
sort of an experiment I wanted to try, and I think the result is that
it's too confusing, and that output parameters are still an awful
pattern whose use cannot possibly be justified.