Implement `hfc status`
This provides a quick overview of currently deployed versions, so it's
easy to see whether all stacks are up to date and in sync.

I reworked the logic for looking up stack tags in clean-uploads, since
the error handling there is much stricter than for getting the status.
If we have any error at all retrieving a stack (doesn't exist, doesn't
have an S3 key, AWS is down), we just say that there's no data instead
of failing the whole status command.