Simplify status command logic
Since a tabwriter obviously can't write anything until it has all of the
output buffered (at least without using the \f escape sequence), it's
obviously not useful to run the tabwriter concurrently with stack
retrieval. Or at least it should be obvious, but apparently wasn't to me
for whatever reason.

Here, we implement the obvious approach of pre-retrieving all stack keys
concurrently, and then just looping over them in the tabwriter output.
In addition to the obvious benefit of simplifying the code, this will
obviously print all of the stacks in a consistent order every time the
command runs, based on the order in which they're defined in the config,
which is obviously a better UX.

Obviously.

Anyways… I also wrote a little wrapper to make tabwriter's error
handling a bit more like bufio.Writer, and to add a couple of
convenience methods so I don't have to look at `[]byte` so much.