Support a global AWS region configuration
I have a particular stack that I can only deploy to us-east-1, where I
have to override the AWS_DEFAULT_REGION environment variable every time
I run hfc since that's not the default region in my local AWS config.
For regular use I have a wrapper script that `go run`s the latest
version of the tool and sets the environment variable automatically, but
when I test new local builds of hfc I always have to remember to do this
myself, and it's a bit annoying.

Well, "was" a bit annoying. Now that this is in hfc itself that workflow
will be a lot easier. Heck, I might just `go install` hfc and stop using
the wrapper script entirely (at least personally, it'll still be in the
repo for reference).