Try out Go's string builder
It's not like the performance benefit here is even remotely meaningful
for our use case... but I was interested in trying out strings.Builder
and I honestly think this version is easier to read.

I guess the secret to strings.Builder is that the final String() call
uses unsafe to transmute an internal byte slice directly into a string,
and has checks to make sure you can't copy builders in ways that would
lead to mutation. Neat.