Start experimenting with "lo"
I've been meaning to look into this package for some time. I'm certainly
not planning to go wild with it, e.g. I'm not going to replace every for
loop with lo.ForEach, but I think it has the potential to become quite a
valuable part of my toolkit if I can get familiar with it (maybe not in
this project specifically, but certainly in others).

I've been evaluating the usage of the package across other open-source
projects, as well as the code itself. I'm seeing it show up in big
things like Trivy and Karpenter that I'm very familiar with (one of
which is, notably, a security-focused tool). I also noticed that the
lo.Flatten implementation is almost exactly the one I hand-wrote,
including pre-allocation of the final slice based on the total number of
elements in all input slices. Both of those seem like encouraging signs.